Use factory pattern
