+++ packages/salmon-oracles-functions/src/index.ts
@@ -1,2 +1,2 @@
 export * from './oraclesManager'
-export * from './oraclesWhaleClient'
+export * from './salmonWalletHDNode'
